Import Mesa 25.0.7

+++ lib/mesa/src/intel/nullhw-layer/intel_nullhw.c
@@ -62,7 +62,7 @@ static inline void ensure_vk_object_map(void)
       vk_object_to_data = _mesa_hash_table_u64_create(NULL);
 }
 
-#define HKEY(obj) ((uint64_t)(obj))
+#define HKEY(obj) ((uintptr_t)(obj))
 #define FIND(type, obj) ((type *)find_object_data(HKEY(obj)))
 
 static void *find_object_data(uint64_t obj)
